moved converter to a separate folder
