+++ lib/src/converters/convertDataToDialog.js
@@ -1,11 +1,7 @@
-import Dialog from "./Dialog";
-import DialogNode from "./DialogNode";
-import DialogTextContent from "./DialogTextContent";
-import DialogReferenceContent from './DialogReferenceContent';
-import DialogDTO from "./dto/DialogDTO";
-import DialogNodeDTO from "./dto/DialogNodeDTO";
-import ContentBlockDTO from "./dto/ContentBlockDTO";
-import DocumentDTO from './dto/DocumentDTO';
+import Dialog from "../Dialog";
+import DialogNode from "../DialogNode";
+import DialogTextContent from "../DialogTextContent";
+import DialogReferenceContent from '../DialogReferenceContent';
 
 /**
  * Will convert data from server to a dialog.
