+++ lib/tests/convertDataToDialog.spec.js
@@ -1,4 +1,4 @@
-import convertDataToDialog from "./convertDataToDialog";
+import convertDataToDialog from "../src/converters/convertDataToDialog";
 import DialogDTO from "./dto/DialogDTO";
 import DialogNodeDTO from "./dto/DialogNodeDTO";
 import ContentBlockDTO from "./dto/ContentBlockDTO";
