+++ lib/tests/dto/ContentBlockDTO.js
