+++ lib/tests/dto/DialogDTO.js
