+++ lib/tests/dto/DialogNodeDTO.js
